Conflicts:
	tools/Cron_jobs/quarantine_maint.php
	tools/Cron_jobs/quarantine_report.php

+++ tools/Cron_jobs/quarantine_maint.php
@@ -85,7 +85,7 @@ if ($required_constant_missing_count == 0) {
                         $sql = "UPDATE maillog SET timestamp=timestamp, quarantined = NULL WHERE id='$id'";
                         dbquery($sql);
                         //If auto quarantine release is enabled, remove from autorelease table when quarantined email expires
-                        if (AUTO_RELEASE) {
+                        if (defined('AUTO_RELEASE') && AUTO_RELEASE === true) {
                             $sql = "DELETE FROM autorelease WHERE msg_id = '$id'";
                             dbquery($sql);
                         }
